+++ fabric21/src/main/java/com/vecoo/extralib/storage/ICustomJsonProvider.java
@@ -1,0 +1,11 @@
+package com.vecoo.extralib.storage;
+
+import java.util.Collection;
+
+public interface ICustomJsonProvider<T> {
+    Collection<T> getStorage();
+
+    void write();
+
+    void init();
+}